Allow Req redirect follow instead of manual
The default for a request has been to pass in redirect manual to fetch
when doing real requests. This is sometimes problematic for an API that
responds with a redirect to another url with the actual content that a
tape should contain instead. With this change the redirect property can
be changed to redirect follow (which is the normal default for fetch).

Usage:

    const request = { url, method, body, headers, redirect: 'follow' }
    requestHandler.handle(request)